requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
greenlet==3.0.3
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyarrow==15.0.0
pyodbc==5.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
